+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - remove requirement for conditionalPredicates to be unique on elementId
+- `excludedCSSClasses` to `pdfSubmissionEventConfiguration`
 
 ## [2.1.0] - 2023-03-26
 
@@ -20,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `hintPosition` property to element validation
 - `toEmail`, `ccEmail` and `bccEmail` to `emailSubmissionEventConfiguration`
 - html string validation to form element hint
-- `excludedCSSClasses` to `pdfSubmissionEventConfiguration`
 
 ### Changed