+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - elementId validation for date/datetime elements
 
+### Changed
+
+- validate elements inside repeatable sets
+
 ## [2.2.0] - 2023-04-02
 
 ### Changed